fix: Avoid spawning a subprocess without a shell
Python possesses many mechanisms to invoke an external executable.
However, doing so may present a security issue if appropriate care is
not taken to sanitize any user provided or variable input. The spawning
of a subprocess without the use of a command shell is not vulnerable to
shell injection attacks, but care should still be taken to ensure
validity of input.

Reference:
https://docs.openstack.org/bandit/latest/plugins/b603_subprocess_without_shell_equals_true.html